% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter-gavi.R
\name{add_gavi_group}
\alias{add_gavi_group}
\title{Add Gavi groups}
\usage{
add_gavi_group(data, ...)
}
\arguments{
\item{data}{A data frame. Required to have a column labeled iso3 with iso3 country codes.}

\item{...}{A group or groups which is in groupings file.}
}
\value{
A data frame with the added groups
}
\description{
This function takes a data frame and adds the specified Gavi groups to it.
}
\examples{

wuenic \%>\%
add_gavi_group(country, who_region)
}
