% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter-gavi.R
\name{filter_gavi_group}
\alias{filter_gavi_group}
\title{Filter by gavi groups}
\usage{
filter_gavi_group(data, group, group_value, drop = TRUE)
}
\arguments{
\item{data}{A data frame. Required to have a column labeled iso3 with iso3 country codes.}

\item{group}{A group which is in groupings file.}

\item{group_value}{A value or vector to filter the data by. Must correspond to values found in the group variable.}

\item{drop}{defaults to TRUE, which drops the filter variable. Use FALSE if you wish to keep the filter varible.}
}
\value{
A filtered data frame.
}
\description{
This function takes a data frame, adds the grouping file and filters it by the specified group and filter values. It then removes the filter group.
}
\examples{

wuenic \%>\%
filter_gavi_group(gavi57, 1)

wuenic \%>\%
filter_gavi_group(who_region, c("AFRO", "EMRO"))

}
