% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gavi-colors.R
\name{gavi_colors}
\alias{gavi_colors}
\title{Gavi colors}
\usage{
gavi_colors(...)
}
\arguments{
\item{...}{a list of gavi color labels}
}
\value{
a character vector with color names as labels and hex numbers as values
}
\description{
Gavi colors
}
