% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-colors.R
\docType{data}
\name{gavi_cols}
\alias{gavi_cols}
\title{Gavi colors}
\format{
A named characted vector with color names as the name and their respective hex number as their value
}
\usage{
gavi_cols
}
\description{
A set of Gavi-defined corportate colors for use in \code{gavi_colors()}
}
\keyword{datasets}
