% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-colors.R
\docType{data}
\name{gavi_pal}
\alias{gavi_pal}
\title{Gavi color palettes}
\format{
A list with 13 named character vectors (each one representing a premade palette)
}
\usage{
gavi_pal
}
\description{
A set of premade color palettes for use in \code{scale_color_gavi()} and \code{scale_fill_gavi()}
}
\keyword{datasets}
