% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gavi-colors.R
\name{gavi_palette}
\alias{gavi_palette}
\title{Gavi palette}
\usage{
gavi_palette(palette = "all", reverse = FALSE, ...)
}
\arguments{
\item{palette}{A name of a Gavi palette (eg. "all").}

\item{reverse}{Reverses order of the palette}

\item{...}{I don't know.}
}
\value{
A character vector with all colors in the named pallette, with names as labels and hex numbers as values.
}
\description{
Gavi palette
}
