% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-data.R
\name{get_groupings}
\alias{get_groupings}
\title{Load in the groupings file}
\usage{
get_groupings(version_date)
}
\arguments{
\item{version_date}{a string in the format of "yyyy-mm", corresponding to the version of the country groupings file you wish to load.}
}
\value{
a data frame of country groupings
}
\description{
This function loads in the country groupings file into your current environment
}
\examples{
get_groupings("2022-07")
# can also save as object to re-use
groupings <- get_groupings("2022-07")

}
