% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-data.R
\name{get_jrf}
\alias{get_jrf}
\title{Retrieve JRF data}
\usage{
get_jrf(version_date, source = "admin", guids = FALSE)
}
\arguments{
\item{version_date}{A string in the format of "yyyy-mm-dd", corresponding to the version of the JRF file you wish to load.}

\item{source}{A string used to determine which format of JRF file to load. Options are: "admin" (default), "official", "system.", and "subnational".}

\item{guids}{Use TRUE if you want a file with GUIDs. Only relevant for subnational data.}
}
\value{
a JRF data frame.
}
\description{
Retrieves JRF data from the shared drive. NOTE: Only works for 2021 or later versions.
}
\examples{

get_jrf("2022-08-03")

get_jrf("2022-09-02", "subnational", guids = TRUE)

# to dynamically always pull most recent file
get_jrf(get_jrf_dates("official", latest = TRUE), "official")
}
