% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-data.R
\name{get_jrf_dates}
\alias{get_jrf_dates}
\title{Retreive dates of JRF files}
\usage{
get_jrf_dates(source = "admin", latest = FALSE, guids = FALSE)
}
\arguments{
\item{source}{A string used to determine which format of JRF file to load. Options are: "admin" (default), "official", "system.", and "subnational".}

\item{latest}{Use TRUE if you only want latest file date.}

\item{guids}{Use TRUE if you want a file with GUIDs. Only relevant for subnational data.}
}
\value{
A vector of dates.
}
\description{
This is a helper function to look at dates for the JRF files.
}
\examples{

get_jrf_dates()
get_jrf_dates("official")
get_jrf_dates("subnational", guids = TRUE)

}
