% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-data.R
\name{get_shapes}
\alias{get_shapes}
\title{Retrieve Shape data}
\usage{
get_shapes(version_date = NULL, source = "polio", layer = "admin2", latest = T)
}
\arguments{
\item{version_date}{A string in the format of "yyyy-mm", corresponding to the version of the shape file you wish to load.}

\item{source}{A string used to determine which shape file to load. Options are: "polio" and "gadm".}

\item{layer}{A string used to determine which layer to load. Options are: "admin0", "admin1", and "admin2"}

\item{latest}{Defaults to TRUE. If loading polio shapes, this ensures you only use the latest shape version for each country
for a given version_date. If set to FALSE, you will load the file which has all shape file versions available for a given version date.}
}
\value{
A sf object
}
\description{
Retrieves Shape data from the shared drive.
}
\examples{

# Defaults are set to polio, admin2, and latest shape version,
# So you can just provide version_date:
get_shapes("2022-03")

# Note for GADM, no version date needed
get_shapes(source = "gadm")

# can also add arguments to override defaults
get_shapes("2022-03", layer = "admin1", latest = FALSE)

}
