% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-data.R
\name{get_support}
\alias{get_support}
\title{Retreive Support data}
\usage{
get_support(version_date)
}
\arguments{
\item{version_date}{A string in the format of "yyyy-mm", corresponding to the version of the support file you wish to load.}
}
\value{
a data frame of country-vaccine-year support
}
\description{
Retreives Support data from the shared drive.
}
\examples{
get_support("2023-08")

}
