% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-data.R
\name{get_survey}
\alias{get_survey}
\title{Retrieve Survey data}
\usage{
get_survey(version_date, source = "national")
}
\arguments{
\item{version_date}{A string in the format of "yyyy-mm", corresponding to the version of the survey file you wish to load.}

\item{source}{A string used to determine which survey file to load. Options are: "national" and "equity".}
}
\value{
a survey data frame.
}
\description{
Retrieves Survey data from the shared drive. NOTE: Only works for 2021 or later versions.
}
\examples{

get_survey("2022-08")

get_survey("2022-08", source = "equity")

}
