% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-data.R
\name{get_synonyms}
\alias{get_synonyms}
\title{Load in the synonyms file}
\usage{
get_synonyms()
}
\value{
a data frame with iso3 and matching country names
}
\description{
This function loads in a simplified version of the synonyms file with just iso3 and country names. Used to match country names with iso3 when no iso3 is in the data.
}
\examples{
get_synonyms()
}
