% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-data.R
\name{get_worldpop}
\alias{get_worldpop}
\title{Retrieve WorldPop data}
\usage{
get_worldpop(version_date, source = "pop")
}
\arguments{
\item{version_date}{A string in the format of "yyyy-mm", corresponding to the version of the WorldPop file you wish to load.}

\item{source}{A string used to determine which WorldPop file to load. Options are: "pop" and "erg_district", "erg_national", "dhf_district", "dhf_national".}
}
\value{
a WorldPop data frame
}
\description{
Retrieves WorldPop data from the shared drive.
}
\examples{

# defaults to population dataset
get_worldpop("2023-07")

# for ERG and distance to facility data sets, indicate if it is district or national level
get_worldpop("2023-07", source = "erg_district")

}
