% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-data.R
\name{get_wpp}
\alias{get_wpp}
\title{Retrieve WPP data}
\usage{
get_wpp(version_date, source = "surviving_infants")
}
\arguments{
\item{version_date}{A string in the format of "yyyy", corresponding to the WPP version you wish to load.}

\item{source}{A string used to determine which WPP file to load. Options are: "surviving_infants", "demographics", and "total_pop".}
}
\value{
a WPP data frame or list (if using the demographics file)
}
\description{
Retrieve WPP data
}
\examples{
get_wpp("2022")
get_wpp("2022", "demographics")

}
