% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-data.R
\name{get_wuenic}
\alias{get_wuenic}
\title{Retrieve WUENIC data}
\usage{
get_wuenic(version_date, format = "normal")
}
\arguments{
\item{version_date}{A string in the format of "yyyy-mm", corresponding to the version of the WUENIC file you wish to load.}

\item{format}{A string used to determine which format of WUENIC file to load. Options are: "normal", "old_new", and "to_share".}
}
\value{
a WUENIC data frame
}
\description{
Retrieves WUENIC data from the shared drive. NOTE: Only works for 2021 or later versions.
}
\examples{

get_wuenic("2022-07")
get_wuenic("2021-07", "old_new")
}
