% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gavi-colors.R
\name{plot_gavi_colors}
\alias{plot_gavi_colors}
\title{Plot Gavi color scales}
\usage{
plot_gavi_colors(pal_name = "all")
}
\arguments{
\item{pal_name}{defaults to all colors. Can choose from a list of available palettes:
all, main, strategy, vaccine, blues, purples, greens, greens3, bluegreen, redgreen, traffic, misc1, misc2}
}
\description{
Helper function to assist in choosing a Gavi color or palette.
}
\examples{
plot_gavi_colors()
plot_gavi_colors("vaccine")

}
