% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-data.R
\name{set_root}
\alias{set_root}
\title{Set root for file paths}
\usage{
set_root(datasets = F)
}
\arguments{
\item{datasets}{set to TRUE if root should be the CPMM datasets folder, otherwise leave as FALSE to have root at CPMM folder.}
}
\value{
a vector with the filepath root.
}
\description{
This function just sets the root directory for filepaths. It simplifies data-loading and dynamically pulls in the username to the filepath.
}
\examples{
set_root()

# can also save as root object to re-use when loading multiple data files
root <- set_root()
}
