% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gavi-themes.R
\name{theme_gavi}
\alias{theme_gavi}
\title{Gavi ggplot2 theme}
\usage{
theme_gavi()
}
\description{
A simple theme based on \code{\link[hrbrthemes:theme_ipsum]{hrbrthemes::theme_ipsum()}} with slight tweaks.
}
