% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gavi-themes.R
\name{theme_gavi_table}
\alias{theme_gavi_table}
\title{Gavi flextable theme}
\usage{
theme_gavi_table(ft)
}
\arguments{
\item{ft}{A flextable object.}
}
\description{
A theme which applies some pretty defaults to a flextable.
}
\examples{

library(dplyr)
library(flextable)
library(officer)

wuenic \%>\%
  filter(year == 2020 & iso3 \%in\% c("COD", "ETH", "IND", "PAK") & vaccine == "dtp3") \%>\%
  select(iso3, coverage) \%>\%
  flextable() \%>\%
  add_header_lines("DTP3 coverage in high impact countries") \%>\%
  add_footer_lines("Source: WUENIC, July 2021") \%>\%
  theme_gavi_table()

}
