% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{wuenic}
\alias{wuenic}
\title{WUENIC Coverage Estimates}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 17424 rows and 6 columns.
}
\source{
WUENIC Estimates, July 2022
}
\usage{
wuenic
}
\description{
A subset of data from the WHO/UNICEF Estimates of National Vaccine Coverage.
Includes variables for country, year, vaccine, target, number vaccinated, and coverage.
Note: Only DTP1, DTP3, MCV1, and PCV3 vaccines are included.
}
\keyword{datasets}
