% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gavi-colors.R
\name{gavi_palette}
\alias{gavi_palette}
\title{Gavi palettes}
\usage{
gavi_palette(pal_name = "all", reverse = FALSE, unnamed = FALSE)
}
\arguments{
\item{pal_name}{A Gavi palette. defaults to all colors. Full list of available palettes: all, standard, millennial, main, strategy, vaccine, who_region, gavi_segment, world_bank, red_green, traffic, misc1, misc2, map_blue_red, map_blue_red2, map_orange_turquoise, map_green_red, map_blues, map_reds, map_greens}

\item{reverse}{Reverses order of the palette.}

\item{unnamed}{Returns an unnamed vector if TRUE.}
}
\value{
A named character vector with all colors and hexcode in the palette.
}
\description{
Gavi palettes
}
\examples{
gavi_palette()
gavi_palette("standard")

}
