% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gavi-colors.R
\name{plot_gavi_colors}
\alias{plot_gavi_colors}
\title{Plot Gavi color scales}
\usage{
plot_gavi_colors(pal_name = "all")
}
\arguments{
\item{pal_name}{A Gavi palette. Defaults to all colors. Full list of available palettes: all, standard, millennial, main, strategy, vaccine, who_region, gavi_segment, world_bank, red_green, traffic, misc1, misc2, map_blue_red, map_blue_red2, map_orange_turquoise, map_green_red, map_blues, map_reds, map_greens}
}
\description{
Helper function to assist in choosing a Gavi color or palette.
}
\examples{
plot_gavi_colors()
plot_gavi_colors("vaccine")

}
