% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gavi-colors.R
\name{scale_fill_gavi}
\alias{scale_fill_gavi}
\title{Fill scales based on Gavi color palette}
\usage{
scale_fill_gavi(palette = "all", discrete = TRUE, reverse = FALSE, ...)
}
\arguments{
\item{palette}{A palette. Defaults to all 21 Gavi colors, but other versions can be used. Full list: all, standard, main, strategy, vaccine, blues, purples, greens, greens3, bluegreen, redgreen, traffic, misc1, misc2, map, map blues.}

\item{discrete}{Defaults to TRUE, keep if data is discrete. If continuous, used FALSE.}

\item{reverse}{Reverses order of palette.}

\item{...}{Additional arguments passed to \code{\link[ggplot2:discrete_scale]{ggplot2::discrete_scale()}} or \code{\link[ggplot2:scale_gradient]{ggplot2::scale_fill_gradientn()}}, used respectively when discrete is TRUE or FALSE}
}
\description{
Fill scales based on Gavi color palette
}
\examples{
# Example one - use main palette as output has four colors needed
library(dplyr)
library(ggplot2)
wuenic \%>\%
 filter(iso3 == "AFG" & year == 2021) \%>\%
 ggplot(aes(vaccine, coverage, fill = vaccine)) +
 geom_bar(stat = "identity") +
 scale_fill_gavi(palette = "main", reverse = TRUE)

# Example two - the "greens" palette has only two colors, so palette interpolates shades
wuenic \%>\%
 filter(iso3 == "AFG" & year == 2021) \%>\%
 ggplot(aes(vaccine, coverage, fill = vaccine)) +
 geom_bar(stat = "identity") +
 scale_fill_gavi(palette = "greens", reverse = TRUE)
}
